requirements.txt
crewai==0.28.8
crewai-tools>=0.1.0
google-generativeai>=0.3.0
jira>=3.4.1
pytest==7.4.3
playwright==1.39.0
pytest-html==4.0.2
browserstack-sdk>=1.0.0
requests>=2.28.0
faker==19.6.2
pyyaml==6.0.1
pytest-xvfb==2.0.0
selenium==4.15.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
aiohttp==3.9.1
asyncio-throttle==1.0.2
pandas==2.1.1
numpy==1.25.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
dash==2.14.1
redis==5.0.1
celery==5.3.4
kombu==5.3.4
prometheus-client==0.18.0
tenacity==8.2.3
backoff==2.2.1
pybreaker==0.8.0
memory-profiler==0.61.0
psutil==5.9.6
gputil==1.4.0
py-cpuinfo==9.0.0
html5lib==1.1
lxml==4.9.3
behave>=1.2.6
robotframework>=4.0.0
python-dotenv>=0.19.0
